# Resolve operator outside loop over value list in expression interpreter