--- a/core/trino-main/src/test/java/io/trino/sql/TestExpressionInterpreter.java
+++ b/core/trino-main/src/test/java/io/trino/sql/TestExpressionInterpreter.java
@@ -1476,8 +1476,11 @@ public class TestExpressionInterpreter
     {
         assertRoundTrip(expression);
 
-        Expression parsedExpression = planExpression(expression);
+        return optimize(planExpression(expression));
+    }
 
+    static Object optimize(Expression parsedExpression)
+    {
         Map<NodeRef<Expression>, Type> expressionTypes = getTypes(TEST_SESSION, METADATA, SYMBOL_TYPES, parsedExpression);
         ExpressionInterpreter interpreter = expressionOptimizer(parsedExpression, METADATA, TEST_SESSION, expressionTypes);
         return interpreter.optimize(symbol -> {
@@ -1513,7 +1516,7 @@ public class TestExpressionInterpreter
     }
 
     // TODO replace that method with io.trino.sql.ExpressionTestUtils.planExpression
-    private static Expression planExpression(@Language("SQL") String expression)
+    static Expression planExpression(@Language("SQL") String expression)
     {
         return TransactionBuilder.transaction(new TestingTransactionManager(), new AllowAllAccessControl())
                 .singleStatement()