Reduce the time to reslove operator for issue #7015