@findepi 2021-02-27
Can you paste BenchmarkExpressionInterpreter from before the change and after?