@lzeiming 2021-03-01
### Before the change
```
Benchmark                                (inValuesCount)  Mode  Cnt     Score     Error  Units
BenchmarkExpressionInterpreter.optimize              200  avgt   20   108.974 ±  68.044  ms/op
BenchmarkExpressionInterpreter.optimize              500  avgt   20   157.350 ±  45.820  ms/op
BenchmarkExpressionInterpreter.optimize             1000  avgt   20   224.354 ±  10.850  ms/op
BenchmarkExpressionInterpreter.optimize             5000  avgt   20  1251.051 ± 113.997  ms/op
```

### After the change
```
Benchmark                                (inValuesCount)  Mode  Cnt    Score    Error  Units
BenchmarkExpressionInterpreter.optimize              200  avgt   20   35.808 ±  2.945  ms/op
BenchmarkExpressionInterpreter.optimize              500  avgt   20   96.284 ±  6.462  ms/op
BenchmarkExpressionInterpreter.optimize             1000  avgt   20  186.406 ± 17.595  ms/op
BenchmarkExpressionInterpreter.optimize             5000  avgt   20  967.069 ± 81.065  ms/op

```