@findepi 2021-03-01
The gain on large inValuesCount is smaller, as if the benefit of the change were diminishing (alleviated by JIT?).

What if we changed this line
https://github.com/trinodb/trino/blob/8fb4a95ab491d91dc82f9a566eaa519ddbc2097a/core/trino-main/src/main/java/io/trino/sql/planner/ExpressionInterpreter.java#L621

to 
```
ResolvedFunction equalsOperator =  metadata.resolveOperator(OperatorType.EQUAL, types(node.getValue(), node.getValue()));
Boolean result = (Boolean) functionInvoker.invoke(equalsOperator, session, ImmutableList.of(value, inValue));
```

(note that I am replacing `types(node.getValue(), expression)` with `types(node.getValue(), node.getValue())` but this should be the same type)

Then you can pull `resolveOperator` before the `for` loop there.

Could you consider such a change and running your benchmark on it?

This can be further combine it with operator cache or not -- if the cache still provides visible gains.