@lzeiming 2021-03-02
### Resolve operator outside loop over value list. Without operator cache change.

```
Benchmark                                (inValuesCount)  Mode  Cnt    Score    Error  Units
BenchmarkExpressionInterpreter.optimize              200  avgt   20   33.364 ±  2.237  ms/op
BenchmarkExpressionInterpreter.optimize              500  avgt   20   91.773 ± 11.335  ms/op
BenchmarkExpressionInterpreter.optimize             1000  avgt   20  154.433 ±  4.217  ms/op
BenchmarkExpressionInterpreter.optimize             5000  avgt   20  812.423 ± 23.894  ms/op
```

### Combining with cache does not improve more.

```
Benchmark                                (inValuesCount)  Mode  Cnt    Score     Error  Units
BenchmarkExpressionInterpreter.optimize              200  avgt   20   32.025 ±   1.025  ms/op
BenchmarkExpressionInterpreter.optimize              500  avgt   20   91.298 ±   9.183  ms/op
BenchmarkExpressionInterpreter.optimize             1000  avgt   20  174.265 ±  11.066  ms/op
BenchmarkExpressionInterpreter.optimize             5000  avgt   20  907.463 ± 108.778  ms/op
```